requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
finnhub-python==2.4.20
pandas==2.2.2
Flask-Cors==4.0.1
Werkzeug==3.0.3
pytest==8.2.2
pytest-mock==3.14.0
